+++ persistent-mongoDB/ChangeLog.md
@@ -1,3 +1,10 @@
+# Changelog for persistent-mongoDB
+
+## 2.9.0 (unreleased)
+
+* Removed deprecated `entityToDocument`. Please use `recordToDocument` instead. [#894](https://github.com/yesodweb/persistent/pull/894)
+* Removed deprecated `multiBsonEq`. Please use `anyBsonEq` instead. [#894](https://github.com/yesodweb/persistent/pull/894)
+
 ## 2.8.0
 
 * Switch from `MonadBaseControl` to `MonadUnliftIO`